% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertToDgCMatrix}
\alias{convertToDgCMatrix}
\title{Convert dense and dgT matrices to dgC matrices}
\usage{
convertToDgCMatrix(x)
}
\arguments{
\item{x}{Matrix, dgTMatrix format or dense format}
}
\value{
dgCMatrix.
}
\description{
Convert dense and dgT matrices to dgC matrices
}
