% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfileSimilarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate cosine similarity between two vectors}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2}
}
\description{
Calculate cosine similarity between two vectors
}
