% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTreeNodeNames.R
\name{findTreeNodeNames}
\alias{findTreeNodeNames}
\title{Find node names of a nested list}
\usage{
findTreeNodeNames(cth)
}
\arguments{
\item{cth}{Cell type hierarchy, as a nested list}
}
\value{
vector of node names
}
\description{
Find node names of a nested list
}
\examples{

cth_list <- list(
  structural = c("endothelial", "fibroblast"),
  lymphoid = list(
    `B-lymphoid` = c("B-cell", "pDC", "plasmablast"),
    `T-lymphoid` = c("NK", "T4", "T8", "Treg")
  ),
  myeloid = c("macrophage", "mast", "mDC", "monocyte", "neutrophil")
)
findTreeNodeNames(cth_list)

}
