% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfileSimilarity.R
\name{plotProfileSimilarity}
\alias{plotProfileSimilarity}
\title{Plots a heatmap of cosine similarity between cell types
in a reference profile}
\usage{
plotProfileSimilarity(full_profiles)
}
\arguments{
\item{full_profiles}{Cell profile matrix, genes x cell types}
}
\description{
Plots a heatmap of cosine similarity between cell types
in a reference profile
}
\examples{
library(InSituType)
data("ioprofiles")
plotProfileSimilarity(ioprofiles)

}
