% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTree.R
\name{printTree}
\alias{printTree}
\title{Prints nested list in an easy to read format}
\usage{
printTree(cth, prefix = "")
}
\arguments{
\item{cth}{Cell type hierarchy, as a nested list}
}
\description{
Prints nested list in an easy to read format
}
\examples{
cth_list <- list(structural = c("endothelial", "fibroblast")
                  , lymphoid = list(`B-lymphoid` = c("B-cell", "pDC", "plasmablast")
                                    , `T-lymphoid` = c("NK", "T4", "T8", "Treg"))
                  , myeloid = c("macrophage", "mast", "mDC", "monocyte", "neutrophil")
                  )
printTree(cth_list)

}
