% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInSituTree.R
\name{runInSituTree}
\alias{runInSituTree}
\title{Tree-style cell typing}
\usage{
runInSituTree(
  x,
  neg,
  full_profiles,
  cth,
  name_for_new_annotation = "topLevel",
  cohort = NULL,
  excluded_genes = c(),
  quantile_absolute_expression_difference_param = 0.5,
  quantile_percent_expression_difference_param = 0.5,
  return_summary_annotation = TRUE
)
}
\arguments{
\item{x}{Counts matrix, cells x genes}

\item{neg}{Vector of mean negative controls for each cell}

\item{full_profiles}{Cell profile matrix, genes x cell types,
for all end point cell types}

\item{cth}{Cell type hierarchy, as a nested list}

\item{name_for_new_annotation}{Used in recursive calls, recommended to leave
at default of 'topLevel'.}

\item{cohort}{Vector of cells' cohort membership}

\item{excluded_genes}{Genes to be excluded during fitting with InSituType}

\item{quantile_absolute_expression_difference_param}{Quantile cutoff for absolute expression.  Default = 0.5.
Both quantile cutoffs must be passed to retain gene.}

\item{quantile_percent_expression_difference_param}{Quantile cutoff for percent expression.  Default = 0.5.
Both quantile cutoffs must be passed to retain gene.}

\item{return_summary_annotation}{option to return a summary of celltypes.
Default is TRUE.}
}
\value{
List of each InSituType result and
a summary of all cell type annotations.
}
\description{
Tree-style cell typing
}
\details{
Any cells which did not have any counts in the genes used for
subclustering will not get a cell type assignment; if this happens in the
first round of annotation, these cells will show 'NA' for all annotation
levels. If this happens at later rounds of annotation, the last assignment
will be cascaded right.
Warnings will be shown if the number of genes chosen is small or if the
counts per cell are low, as both can reduce confidence in the cell type
assignments.
}
\examples{

test_celltype_relationship_list <- list(
  "structural" =
    c("endothelial", "fibroblast"),
  "myeloid" = c("macrophage", "mast", "mDC", "monocyte", "neutrophil"),
  "lymphoid" = list(
    "B-lymphoid" = c("B-cell", "pDC", "plasmablast"),
    "T-lymphoid" = list(
      "T4" = c("T CD4 memory", "T CD4 naive", "Treg"),
      "T8" = c("T CD8 memory", "T CD8 naive"),
      "NK"
    )
  )
)

# Extract expression data from insitutype package
library(InSituType)
data(mini_nsclc)
data(ioprofiles)


res <- runInSituTree(
  full_profiles = ioprofiles,
  cth = test_celltype_relationship_list,
  x = as(mini_nsclc$counts, "CsparseMatrix"),
  neg = Matrix::rowMeans(mini_nsclc$neg),
  name_for_new_annotation = "test",
  cohort = NULL,
  excluded_genes = c(
    "MALAT1", "B2M", "CD298",
    "MZT2A", "HLA-A", "HLA-B", "HLA-C"
  )
)

}
