% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedSubcluster.R
\name{supervisedSubcluster}
\alias{supervisedSubcluster}
\title{Supervised subclustering using a gene subset and insitutypeML}
\usage{
supervisedSubcluster(
  reference_profiles,
  x,
  neg,
  quantile_absolute_expression_difference = 0.5,
  quantile_percent_expression_difference = 0.5,
  excluded_genes = NULL,
  cohort = NULL
)
}
\arguments{
\item{reference_profiles}{Matrix of expression profiles, genes x cell types}

\item{x}{Counts matrix, cells x genes}

\item{neg}{Vector of mean negative controls for each cell}

\item{quantile_absolute_expression_difference}{Quantile cutoff for absolute expression.  Default = 0.5. Both quantile cutoffs must be passed to retain gene.}

\item{quantile_percent_expression_difference}{Quantile cutoff for percent expression.  Default = 0.5. Both quantile cutoffs must be passed to retain gene.}

\item{excluded_genes}{Genes to be excluded during fitting with InSituType}

\item{cohort}{Vector of cells' cohort membership}
}
\value{
Insitutype result list.
}
\description{
Supervised subclustering using a gene subset and insitutypeML
}
