% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseProfiles.R
\name{collapseProfiles}
\alias{collapseProfiles}
\title{Collapse a hierarchy of cell expression profiles to the root level
as described in the nested list for cell type hierarchy}
\usage{
collapseProfiles(cth, full_profiles, verbose = FALSE, print_tree = FALSE)
}
\arguments{
\item{cth}{Cell type hierarchy, as a nested list with end point cell types
in full_profiles}

\item{full_profiles}{Cell profile matrix, genes x cell types,
include all end point cell types in cth}

\item{verbose}{Option to print progress to console}

\item{print_tree}{Option to print a structure of the nested lsit}
}
\value{
Matrix of collapsed cell expression profiles
}
\description{
Collapse a hierarchy of cell expression profiles to the root level
as described in the nested list for cell type hierarchy
}
\examples{

set.seed(123)
test_profiles <- structure(
  rnorm(42, mean = 3, sd = 1),
  .Dim = c(3L, 14L),
  .Dimnames = list(
    c("GeneA", "GeneB", "GeneC"),
    c(
      "B-cell", "endothelial", "fibroblast",
      "macrophage", "mast", "mDC", "monocyte", "neutrophil", "NK",
      "pDC", "plasmablast", "T4", "T8", "Treg"
    )
  )
)

cth_list <- list(
  structural = c("endothelial", "fibroblast"),
  lymphoid = list(
    `B-lymphoid` = c(
      "B-cell", "pDC",
      "plasmablast"
    ),
    `T-lymphoid` = c(
      "NK", "T4", "T8",
      "Treg"
    )
  ),
  myeloid = c(
    "macrophage", "mast", "mDC", "monocyte",
    "neutrophil"
  )
)

test <- collapseProfiles(cth = cth_list, full_profiles = test_profiles)
test
}
