% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeInSituTree.R
\name{summarizeInSituTree}
\alias{summarizeInSituTree}
\title{Summarize the annotations for runInSituTree() function}
\usage{
summarizeInSituTree(nestList, annotationLevel = 1)
}
\arguments{
\item{nestList}{Nested insitutype objects returned from runInSituType()}

\item{annotationLevel}{Used for recursive looping.  Do not adjust.}
}
\value{
data.frame of celltype annotation and posterior probabilities
}
\description{
Summarize the annotations for runInSituTree() function
}
