% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedSubcluster.R
\name{supervisedSubcluster}
\alias{supervisedSubcluster}
\title{Supervised subclustering using a gene subset and insitutypeML}
\usage{
supervisedSubcluster(
  reference_profiles,
  x,
  neg,
  quantile_absolute_expression_difference = 0.5,
  quantile_percent_expression_difference = 0.5,
  excluded_genes = NULL,
  cohort = NULL,
  ...
)
}
\arguments{
\item{reference_profiles}{Matrix of expression profiles, genes x cell types}

\item{x}{Counts matrix, cells x genes}

\item{neg}{Vector of mean negative controls for each cell}

\item{quantile_absolute_expression_difference}{Minimum absolute expression
difference within reference_profiles in terms of quantile level among all
genes. Default = 0.5. Both quantile cutoffs must be passed to retain gene.}

\item{quantile_percent_expression_difference}{Minimum percentage expression
difference within reference_profiles in terms of quantile level among all
genes. Default = 0.5. Both quantile cutoffs must be passed to retain gene.}

\item{excluded_genes}{Genes to be excluded during fitting with InSituType}

\item{cohort}{Vector of cells' cohort membership}

\item{...}{Additional parameters passed on to \code{insitutypeML}}
}
\value{
a list with all elements returned by InSituType::insitutypeML() and
additional element \code{ctsPerCell} for a vector of counts per cell within the
chosen subset of genes.
}
\description{
Supervised subclustering using a gene subset and insitutypeML
}
